* commit '25c17ff913a2d5ad894fe938f1a0cef43abff4b1': (248 commits)
  Add --no-implicit-cache-flush option for sys:setup:run command, closes #850
  Improve test messages
  Switch to database-helper in db:dump command
  Polish
  Track changes
  Add command eav:attribute:create-dummy-values (with tests)
  Add command category:create:dummy (with tests)
  PHP 5.3 backwards compatibility
  Share array functions from Magerun 1
  Add shared src files to phar build
  Switch to shared build script, towards #494
  git subrepo pull shared
  Remove old files of new shared ones, towards #494
  Add shared tests to testsuite, towards #494
  git subrepo pull shared
  Switch to shared bootstrap class, towards #494
  git subrepo pull shared
  Add PSR-4 autloader for shared folder
  git subrepo clone ../mr-shared shared
  Add keyed access to config getter
  ...

# Conflicts:
#	.travis.yml
#	build.sh

+++ src/N98/Magento/Command/System/Setup/RemoveCommand.php
@@ -3,14 +3,15 @@
 namespace N98\Magento\Command\System\Setup;
 
 use InvalidArgumentException;
+use Mage_Core_Model_Resource;
+use RuntimeException;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use N98\Util\Console\Helper\Table\Renderer\RendererFactory;
 
 /**
  * Class RemoveCommand
+ *
  * @package N98\Magento\Command\System\Setup
  * @author Aydin Hassan <aydin@hotmail.co.uk>
  */
@@ -26,11 +27,10 @@ class RemoveCommand extends AbstractSetupCommand
             ->addArgument('module', InputArgument::REQUIRED, 'Module name')
             ->addArgument('setup', InputArgument::OPTIONAL, 'Setup code to remove', 'all')
             ->setDescription('Remove module setup resource entry');
-
     }
 
     /**
-     * @param InputInterface  $input
+     * @param InputInterface $input
      * @param OutputInterface $output
      *
      * @return void
@@ -39,25 +39,28 @@ class RemoveCommand extends AbstractSetupCommand
     {
         $this->detectMagento($output, true);
 
-        if ($this->initMagento()) {
-            $moduleName     = $this->getModule($input);
-            $setupName      = $input->getArgument('setup');
-            $moduleSetups   = $this->getModuleSetupResources($moduleName);
+        if (!$this->initMagento()) {
+            return;
+        }
 
-            if (empty($moduleSetups)) {
-                $output->writeln(sprintf('No setup resources found for module: "%s"', $moduleName));
-                return;
-            }
+        $moduleName = $this->getModule($input);
+        $setupName = $input->getArgument('setup');
+        $moduleSetups = $this->getModuleSetupResources($moduleName);
 
-            if ($setupName === 'all') {
-                foreach ($moduleSetups as $setupCode => $setup) {
-                    $this->removeSetupResource($moduleName, $setupCode, $output);
-                }
-            } elseif (array_key_exists($setupName, $moduleSetups)) {
-                $this->removeSetupResource($moduleName, $setupName, $output);
-            } else {
-                throw new InvalidArgumentException(sprintf('Error no setup found with the name: "%s"', $setupName));
+        if (empty($moduleSetups)) {
+            $output->writeln(sprintf('No setup resources found for module: "%s"', $moduleName));
+
+            return;
+        }
+
+        if ($setupName === 'all') {
+            foreach ($moduleSetups as $setupCode => $setup) {
+                $this->removeSetupResource($moduleName, $setupCode, $output);
             }
+        } elseif (array_key_exists($setupName, $moduleSetups)) {
+            $this->removeSetupResource($moduleName, $setupName, $output);
+        } else {
+            throw new InvalidArgumentException(sprintf('Error no setup found with the name: "%s"', $setupName));
         }
     }
 
@@ -69,9 +72,13 @@ class RemoveCommand extends AbstractSetupCommand
      */
     public function removeSetupResource($moduleName, $setupResource, OutputInterface $output)
     {
-        $model          = $this->_getModel('core/resource', 'Mage_Core_Model_Resource');
-        $table          = $model->getTableName('core_resource');
-        $writeAdapter   = $model->getConnection('core_write');
+        /** @var Mage_Core_Model_Resource $model */
+        $model = $this->_getModel('core/resource', 'Mage_Core_Model_Resource');
+        $writeAdapter = $model->getConnection('core_write');
+        if (!$writeAdapter) {
+            throw new RuntimeException('Database not configured');
+        }
+        $table = $model->getTableName('core_resource');
 
         if ($writeAdapter->delete($table, array('code = ?' => $setupResource)) > 0) {
             $output->writeln(
